backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
minio
boto3
python-multipart
tiktoken
pillow
easyocr
openai==0.28.1
pydantic==1.*
pdfplumber
ffmpeg-python
whisper-openai
